+++ src/WindowsPackageManager/Source.def
@@ -5,3 +5,6 @@ EXPORTS
     WindowsPackageManagerServerModuleCreate
     WindowsPackageManagerServerModuleRegister
     WindowsPackageManagerServerModuleUnregister
+    WindowsPackageManagerInProcModuleInitialize
+    WindowsPackageManagerInProcModuleTerminate
+    WindowsPackageManagerInProcModuleGetClassObject
